[REF] split JS file into 3 files, following Odoo conventions
